22107.201_beta1
New rule for naming commits: I will now be dropping the "commits" and leave only the version string. This is because I don't usually commit until I publish a version. Testings are done locally and thus I don't have to push commits every time I make changes.

This release deprecates PowerShell 7 as the main orchestrator. It will now be limited to the music player module (to help installing the PS module on 1507 less bitter). The release also introduces the "Hikaru-chan" mechanism, with the first pieces of it being in this beta.