22107.400_beta7c
To any of you who were scanning the Commits history of this repository: Yes, this commit was where all of the "nonsense" codes that were pointing to variables nowhere to be found (like $ds) were removed. These are actually Easter Egg codes and for the sake of transparency, I decided to no longer include them with the source code. Binary (.exe) releases will still have them included. As for what the Easter Eggs were... perhaps you shouldn't pay attention to it as it's not made for you, the general public.